fix bug in mac

+++ Source/Eto.Mac/Forms/Controls/GridViewHandler.cs
@@ -194,9 +194,9 @@ namespace Eto.Mac.Forms.Controls
 					Control.DoubleClick += (sender, e) =>
 					{
 						int rowIndex;
-						if ((rowIndex = Control.ClickedRow) >= 0)
+						if ((rowIndex = (int)Control.ClickedRow) >= 0)
 						{
-							var columnIndex = Control.ClickedColumn;
+							var columnIndex = (int)Control.ClickedColumn;
 							var item = GetItem(rowIndex);
 							var column = columnIndex == -1 ? null : Widget.Columns[columnIndex];
 							Callback.OnCellDoubleClick(Widget, new GridViewCellEventArgs(column, rowIndex, columnIndex, item));
